backend/requirements.txt
amqp==1.4.9
anyjson==0.3.3
awscli==1.7.41
awsebcli==3.7.7
bcdoc==0.16.0
billiard==3.3.0.23
blessed==1.9.5
boto==2.41.0
botocore==1.4.37
cement==2.8.2
colorama==0.3.7
defusedxml==0.4.1
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.11.6
django-appconf==1.0.1
django-argonauts==1.1.3
django-bootstrap3==9.0.0
django-compressor==1.5
django-cors-headers==2.1.0
django-crispy-forms==1.6.1
django-debug-toolbar==1.4
django-filter==1.0.4
django-model-utils==2.5
django-tinymce==2.6.0
djangorestframework==3.7.0
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
drf-nested-routers==0.9.0
ecdsa==0.13
jmespath==0.9.0
kombu==3.0.35
langdetect==1.0.5
olefile==0.44
paramiko==1.17.1
pathspec==0.3.4
Pillow==4.3.0
prettytable==0.7.2
psycopg2==2.7
pyasn1==0.1.9
pycrypto==2.6.1
python-dateutil==2.5.3
pytz==2017.2
PyYAML==3.12
requests==2.9.1
rsa==3.1.4
semantic-version==2.5.0
six==1.10.0
sqlparse==0.1.19
ssh==1.8.0
static3==0.6.1
texttable==0.8.4
wcwidth==0.1.7
websocket-client==0.37.0
